adjupi_api/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
paho-mqtt==1.5.0
pyrsistent==0.15.7
pytz==2019.3
six==1.14.0
Werkzeug==1.0.0
